# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Tenant Ledger & Financial Analytics API
  version: 1.0.0
  description: |
    Public contract of the ledger platform, served through the API gateway.

    **Authentication.** Every `/api/**` call needs an OAuth2 access token issued by Keycloak
    (realm `ledger`, audience `ledger-api`), sent as `Authorization: Bearer <token>`.

    **Tenancy.** Tenant-scoped endpoints require `X-Tenant-ID: <tenant slug>`. The slug must
    appear in the token's `tenants` claim (Keycloak group `/tenants/<slug>/<ROLE>`); the role
    in that group decides what the caller may do:

    | Role       | Read | Post transactions | Manage accounts | AI queries |
    |------------|------|-------------------|-----------------|------------|
    | OWNER      | yes  | yes               | yes             | yes        |
    | ADMIN      | yes  | yes               | yes             | yes        |
    | ACCOUNTANT | yes  | yes               | yes             | yes        |
    | AUDITOR    | yes  | no                | no              | yes        |
    | VIEWER     | yes  | no                | no              | no         |

    Admin endpoints (`/api/v1/admin/**`) take no tenant header and need the realm role
    `PLATFORM_ADMIN`.

    **Errors.** All errors are RFC 9457 `application/problem+json` with a stable `code`
    (see `ErrorCode`) and a `retryable` flag. Clients branch on `code`, never on `title`/`detail`.

...

...
  contact:
    name: Ledger Platform
  license:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /api/v1/ledger/transaction:
    post:
      tags: [Ledger]
      operationId: postTransaction
      summary: Post a double-entry transaction
      description: |
        Atomically posts a journal entry. For each currency, the sum of DEBIT lines must equal
        the sum of CREDIT lines. Posting is idempotent: repeating a request with the same
        `Idempotency-Key` and the same body returns the original entry (`200` with
        `Idempotent-Replayed: true`); reusing the key with a different body is a `409`.

        Each posted entry is sealed into the tenant's SHA-256 hash chain (`chainSeq`,
        `entryHash`) and cannot be modified afterwards; corrections are new reversing entries.
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostTransactionRequest'
            examples:
              cloudInvoice:
                summary: AWS invoice paid from operating cash
                value:
                  effectiveDate: '2026-10-08'
                  description: AWS invoice 2026-09
                  externalRef: INV-AWS-2026-09
                  metadata: { vendor: aws, costCenter: platform }
                  lines:
                    - { accountCode: '6100', direction: DEBIT, amount: '1250.00', currency: USD, memo: EC2 + RDS }
                    - { accountCode: '1000', direction: CREDIT, amount: '1250.00', currency: USD }
      responses:
        '201':
          description: Transaction posted
          headers:
            Location:
              schema: { type: string, example: /api/v1/ledger/transaction/5ece6ef5-c230-4a41-bbe0-22d0792ca4b1 }
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
            X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '200':
          description: Idempotent replay of an already-posted transaction
          headers:
            Idempotent-Replayed:
              schema: { type: string, enum: ['true'] }
            X-Request-Id: { $ref: '#/components/headers/RequestId' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/TenantNotFound'
        '409':
          description: Idempotency key reused with a different body, or the period is closed
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              examples:
                keyReused:
                  value:
                    type: https://ledger.dev/problems/idempotency-key-reused
                    title: Idempotency-Key was already used for a different request
                    status: 409
                    detail: Idempotency-Key was already used for a different transaction
                    instance: /api/v1/ledger/transaction
                    code: IDEMPOTENCY_KEY_REUSED
                    retryable: false
                    requestId: f61696b1-8cf1-4a3c-87c3-a1169434dd6e
                    tenant: acme
                    transactionId: 5ece6ef5-c230-4a41-bbe0-22d0792ca4b1
        '422':
          description: Business rule violation (unbalanced, unknown account, insufficient funds, ...)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
              examples:
                unbalanced:
                  value:
                    type: https://ledger.dev/problems/ledger-unbalanced
                    title: Debits and credits do not balance
                    status: 422
                    detail: Debits and credits do not balance
                    instance: /api/v1/ledger/transaction
                    code: LEDGER_UNBALANCED
                    retryable: false
                    requestId: 0480f010-8a16-479a-926d-8ac2fc14f4bc
                    tenant: acme
                    errors: [{ field: postTransactionRequest, message: debits and credits differ by 1 USD }]
                insufficientFunds:
                  value:
                    type: https://ledger.dev/problems/insufficient-funds
                    title: Insufficient funds
                    status: 422
                    code: INSUFFICIENT_FUNDS
                    retryable: false
                unknownAccount:
                  value:
                    type: https://ledger.dev/problems/account-not-found
                    title: Unknown account
                    status: 422
                    detail: 'Unknown account(s): 6000'
                    code: ACCOUNT_NOT_FOUND
                    retryable: false
                    accountCodes: ['6000']
        '423':
          $ref: '#/components/responses/TenantUnavailable'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          description: Tenant posting concurrency exhausted or lock contention; safe to retry with the same Idempotency-Key
          headers:
            Retry-After: { $ref: '#/components/headers/RetryAfter' }
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/ledger/transaction/{id}:
    get:
      tags: [Ledger]
      operationId: getTransaction
      summary: Fetch a posted transaction
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: The transaction
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionResponse' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404':
          description: No such transaction in this tenant (code `TRANSACTION_NOT_FOUND`)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/ledger/accounts:
    get:
      tags: [Ledger]
      operationId: listAccounts
      summary: Chart of accounts with current balances
      parameters:
        - $ref: '#/components/parameters/TenantId'
      responses:
        '200':
          description: All accounts of the tenant
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Account' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Ledger]
      operationId: createAccount
      summary: Create an account
      parameters:
        - $ref: '#/components/parameters/TenantId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAccountRequest' }
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Account' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Account code already exists (code `ACCOUNT_CODE_EXISTS`)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/v1/admin/tenants:
    get:
      tags: [Admin]
      operationId: listTenants
      summary: List tenants
      responses:
        '200':
          description: All tenants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tenant' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      operationId: provisionTenant
      summary: Provision a tenant
      description: |
        Registers the tenant, creates its schema (in the shared cluster for `SHARED`, or in the
        supplied empty database for `ISOLATED`), runs the tenant migrations, optionally seeds a
        default chart of accounts, and activates it. A tenant left in `PROVISIONING` by a crash
        is resumed by repeating the call.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProvisionTenantRequest' }
            examples:
              shared:
                value: { slug: initech, displayName: Initech, tier: SHARED, baseCurrency: USD }
              isolated:
                value:
                  slug: umbrella
                  displayName: Umbrella Corp
                  tier: ISOLATED
                  baseCurrency: EUR
                  isolatedDatabase:
                    jdbcUrl: jdbc:postgresql://db-umbrella:5432/ledger_umbrella
                    username: ledger
                    secretRef: env:UMBRELLA_DB_PASSWORD
                    poolMaxSize: 10
      responses:
        '201':
          description: Tenant provisioned and active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Tenant already exists (`TENANT_EXISTS`)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500':
          description: Provisioning failed (`TENANT_PROVISIONING_FAILED`); safe to retry
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/admin/tenants/{slug}:
    get:
      tags: [Admin]
      operationId: getTenant
      summary: Get a tenant
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: The tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '404': { $ref: '#/components/responses/TenantNotFound' }
  /api/v1/admin/tenants/{slug}/suspend:
    post:
      tags: [Admin]
      operationId: suspendTenant
      summary: Suspend a tenant (its API calls then fail with 423)
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Suspended
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '404': { $ref: '#/components/responses/TenantNotFound' }
  /api/v1/admin/tenants/{slug}/activate:
    post:
      tags: [Admin]
      operationId: activateTenant
      summary: Re-activate a suspended tenant
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Active
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '404': { $ref: '#/components/responses/TenantNotFound' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token (realm `ledger`, audience `ledger-api`)
  parameters:
    TenantId:
      name: X-Tenant-ID
      in: header
      required: true
      description: Tenant slug; must be one of the caller's tenants (token `tenants` claim)
      schema: { type: string, pattern: '^[a-z][a-z0-9_]{2,47}$', example: acme }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Client-generated unique key per logical transaction (e.g. a UUID); retries must reuse it
      schema: { type: string, pattern: '^[A-Za-z0-9._:-]{8,128}$', example: 0b9c2a3e-6f1d-4d39-9a7e-2f4f0b3c9d11 }
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Correlation id; generated by the gateway when absent and echoed on the response
      schema: { type: string, pattern: '^[A-Za-z0-9._-]{8,64}$' }
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z][a-z0-9_]{2,47}$' }
  headers:
    RequestId:
      description: Correlation id of this request
      schema: { type: string }
    RateLimitLimit:
      description: Bucket capacity for this tenant and endpoint class
      schema: { type: integer }
    RateLimitRemaining:
      description: Requests left in the bucket
      schema: { type: integer }
    RetryAfter:
      description: Seconds to wait before retrying
      schema: { type: integer }
  responses:
    BadRequest:
      description: Malformed request or validation failure
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            missingTenant:
              value:
                type: https://ledger.dev/problems/tenant-header-missing
                title: X-Tenant-ID header is required
                status: 400
                code: TENANT_HEADER_MISSING
                retryable: false
            validation:
              value:
                type: https://ledger.dev/problems/validation-failed
                title: Request validation failed
                status: 400
                code: VALIDATION_FAILED
                retryable: false
                errors: [{ field: 'lines[0].amount', message: must be greater than 0 }]
    Unauthenticated:
      description: Missing, expired or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Caller is not a member of the tenant (`TENANT_ACCESS_DENIED`) or its role lacks the permission (`PERMISSION_DENIED`)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: https://ledger.dev/problems/tenant-access-denied
            title: No access to this tenant
            status: 403
            detail: You are not a member of tenant umbrella
            code: TENANT_ACCESS_DENIED
            retryable: false
    TenantNotFound:
      description: Tenant does not exist (`TENANT_NOT_FOUND`)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    TenantUnavailable:
      description: Tenant is suspended or still provisioning (`TENANT_UNAVAILABLE`)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    RateLimited:
      description: Per-tenant rate limit exceeded (`RATE_LIMITED`)
      headers:
        Retry-After: { $ref: '#/components/headers/RetryAfter' }
        X-RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
        X-RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Money:
      description: Decimal amount, up to 16 integer and 4 fractional digits
      type: string
      pattern: '^\d{1,16}(\.\d{1,4})?$'
      example: '1250.00'
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      example: USD
    Direction:
      type: string
      enum: [DEBIT, CREDIT]
    AccountType:
      type: string
      enum: [ASSET, LIABILITY, EQUITY, REVENUE, EXPENSE]
    PostTransactionRequest:
      type: object
      required: [effectiveDate, description, lines]
      properties:
        effectiveDate:
          type: string
          format: date
          description: Accounting date; its period must be open
        description: { type: string, maxLength: 500 }
        externalRef: { type: string, maxLength: 128, description: 'Reference in the source system, e.g. invoice number or bank transaction id' }
        metadata:
          type: object
          maxProperties: 50
          additionalProperties: true
        lines:
          type: array
          minItems: 2
          maxItems: 500
          description: Per currency, sum(DEBIT amounts) must equal sum(CREDIT amounts)
          items: { $ref: '#/components/schemas/PostingLine' }
    PostingLine:
      type: object
      required: [accountCode, direction, amount, currency]
      properties:
        accountCode: { type: string, maxLength: 32, example: '6100' }
        direction: { $ref: '#/components/schemas/Direction' }
        amount:
          allOf: [{ $ref: '#/components/schemas/Money' }]
          description: Strictly positive; the sign comes from `direction`
        currency:
          allOf: [{ $ref: '#/components/schemas/Currency' }]
          description: Must equal the account's currency
        memo: { type: string, maxLength: 500 }
    TransactionResponse:
      type: object
      required: [id, entryNo, idempotencyKey, effectiveDate, postedAt, description, source, lines, totals]
      properties:
        id: { type: string, format: uuid }
        entryNo: { type: integer, format: int64, description: Gap-free per-tenant sequence number }
        idempotencyKey: { type: string }
        effectiveDate: { type: string, format: date }
        postedAt: { type: string, format: date-time }
        description: { type: string }
        source: { type: string, enum: [API, SEC_EDGAR, PAYSIM, RECONCILIATION, REVERSAL, SYSTEM], example: API }
        externalRef: { type: string, nullable: true }
        metadata: { type: object, additionalProperties: true }
        lines:
          type: array
          items:
            type: object
            properties:
              lineNo: { type: integer }
              accountCode: { type: string }
              accountName: { type: string }
              direction: { $ref: '#/components/schemas/Direction' }
              amount: { type: number, example: 1250.0 }
              currency: { $ref: '#/components/schemas/Currency' }
              memo: { type: string, nullable: true }
        totals:
          type: array
          items:
            type: object
            properties:
              currency: { $ref: '#/components/schemas/Currency' }
              debits: { type: number }
              credits: { type: number }
        chainSeq:
          type: integer
          format: int64
          nullable: true
          description: Position in the tenant's tamper-evident hash chain
        entryHash:
          type: string
          nullable: true
          description: Hex SHA-256 of this entry chained to the previous one
          example: 9f2c4e0d7b1a...
    Account:
      type: object
      properties:
        id: { type: integer, format: int64 }
        code: { type: string, example: '1000' }
        name: { type: string, example: Operating Cash }
        type: { $ref: '#/components/schemas/AccountType' }
        normalBalance: { $ref: '#/components/schemas/Direction' }
        currency: { $ref: '#/components/schemas/Currency' }
        category: { type: string, nullable: true, example: CLOUD }
        allowNegative: { type: boolean, description: 'When false, postings that would overdraw the account fail with INSUFFICIENT_FUNDS' }
        active: { type: boolean }
        parentCode: { type: string, nullable: true }
        debitTotal: { type: number }
        creditTotal: { type: number }
        balance: { type: number, description: Balance in the account's normal direction (positive = normal) }
    CreateAccountRequest:
      type: object
      required: [code, name, type, currency]
      properties:
        code: { type: string, maxLength: 32, pattern: '^[A-Za-z0-9._-]+$' }
        name: { type: string, maxLength: 200 }
        type: { $ref: '#/components/schemas/AccountType' }
        normalBalance:
          allOf: [{ $ref: '#/components/schemas/Direction' }]
          description: Defaults to DEBIT for assets/expenses, CREDIT otherwise
        currency: { $ref: '#/components/schemas/Currency' }
        category: { type: string, maxLength: 64 }
        allowNegative: { type: boolean }
        parentCode: { type: string, maxLength: 32 }
        externalRef: { type: string, maxLength: 128 }
...
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        displayName: { type: string }
        tier: { type: string, enum: [SHARED, ISOLATED] }
        status: { type: string, enum: [PROVISIONING, ACTIVE, SUSPENDED, OFFBOARDED] }
        schemaName: { type: string, example: t_acme }
        baseCurrency: { $ref: '#/components/schemas/Currency' }
    ProvisionTenantRequest:
      type: object
      required: [slug, displayName]
      properties:
        slug: { type: string, pattern: '^[a-z][a-z0-9_]{2,47}$' }
        displayName: { type: string, maxLength: 200 }
        tier: { type: string, enum: [SHARED, ISOLATED], default: SHARED }
        baseCurrency: { allOf: [{ $ref: '#/components/schemas/Currency' }], default: USD }
        seedChartOfAccounts: { type: boolean, default: true }
        isolatedDatabase:
          type: object
          description: Required for ISOLATED, forbidden for SHARED. Points at an existing, empty database.
          required: [jdbcUrl, username, secretRef]
          properties:
            jdbcUrl: { type: string, pattern: '^jdbc:postgresql://.+' }
            username: { type: string, maxLength: 63 }
            secretRef: { type: string, pattern: '^(env|file):.+', description: Where the password is read from; never the password itself }
            poolMaxSize: { type: integer, minimum: 1, maximum: 200, default: 10 }
    Problem:
      type: object
      description: RFC 9457 problem details with ledger extensions
      required: [type, title, status, code, retryable]
      properties:
        type: { type: string, format: uri, example: https://ledger.dev/problems/ledger-unbalanced }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { $ref: '#/components/schemas/ErrorCode' }
        retryable: { type: boolean, description: Whether the same request may succeed if retried (with the same Idempotency-Key) }
        requestId: { type: string }
        tenant: { type: string }
        errors:
          type: array
          description: Field-level validation errors
          items:
            type: object
            properties:
              field: { type: string }
              message: { type: string }
      additionalProperties: true
    ErrorCode:
      type: string
      enum:
        - VALIDATION_FAILED
        - MALFORMED_REQUEST
        - TENANT_HEADER_MISSING
        - TENANT_HEADER_INVALID
        - IDEMPOTENCY_KEY_MISSING
        - IDEMPOTENCY_KEY_INVALID
        - UNAUTHENTICATED
        - TENANT_ACCESS_DENIED
        - PERMISSION_DENIED
        - TENANT_NOT_FOUND
        - TRANSACTION_NOT_FOUND
        - RESOURCE_NOT_FOUND
        - IDEMPOTENCY_KEY_REUSED
        - ACCOUNT_CODE_EXISTS
        - TENANT_EXISTS
        - JOURNAL_SEALED
        - LEDGER_IMMUTABLE
        - PERIOD_CLOSED
        - LEDGER_UNBALANCED
        - INSUFFICIENT_LINES
        - ACCOUNT_NOT_FOUND
        - ACCOUNT_INACTIVE
        - CURRENCY_MISMATCH
        - INSUFFICIENT_FUNDS
        - PERIOD_NOT_OPEN
        - TENANT_UNAVAILABLE
        - RATE_LIMITED
        - AI_BUDGET_EXCEEDED
        - AI_PROVIDER_UNAVAILABLE
        - TENANT_BUSY
        - LEDGER_CONTENTION
        - TENANT_PROVISIONING_FAILED
        - INTERNAL_ERROR